# ec_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/exercises/query": {
      "get": {
        "operationId": "ExerciseController_fetchExercisesByQuery",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Level",
              "example": "A1",
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Difficulty",
              "example": "EASY",
              "type": "string"
            }
          },
          {
            "name": "quizId",
            "required": false,
            "in": "query",
            "schema": {
              "title": "QuizID",
              "example": "3db18797-e514-40ac-9a2e-52927016476d",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The exercise has been successfully fetched."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The exercise was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ]
      }
    },
    "/api/exercises/{id}": {
      "get": {
        "operationId": "ExerciseController_fetchExerciseById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The exercise has been successfully fetched."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The exercise was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ]
      },
      "patch": {
        "operationId": "ExerciseController_updateExercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The exercise has been successfully updated!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The exercise was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ]
      },
      "delete": {
        "operationId": "ExerciseController_deleteExercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The exercise has been deleted."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (numeric string is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The exercise was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ]
      }
    },
...
    "/api/answers/e": {
      "post": {
        "operationId": "AnsweredExerciseController_saveAnswer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAnswerDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answer has been successfully saved."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The answeredQuizId is not a valid UUID."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      },
      "get": {
        "operationId": "AnsweredExerciseController_fetchAnswers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answers have been successfully fetched."
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "There are no answers to show."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      }
    },
    "/api/answers/e/query": {
      "get": {
        "operationId": "AnsweredExerciseController_fetchAnswersByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answers were successfully fetched."
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "There are no answers to show or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      }
    },
    "/api/answers/e/{id}": {
      "get": {
        "operationId": "AnsweredExerciseController_fetchAnswerById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answer was successfully fetched."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The answer was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      },
      "patch": {
        "operationId": "AnsweredExerciseController_addFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answer was successfully updated."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The answer was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      },
      "delete": {
        "operationId": "AnsweredExerciseController_deleteAnswer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The answer was successfully deleted."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": "The answer was not found or the ID is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "AnsweredExercise"
        ]
      }
    },
...
  },
  "info": {
    "title": "English Conversations API",
    "description": "REST API for managing English Conversations' data and processes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SaveAnswerDTO": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "number",
            "title": "ExerciseID",
            "example": 1
          },
          "userId": {
            "type": "string",
            "title": "UserID",
            "example": "fa7f094c-be3a-4c1f-9369-957f7d91d05b"
          },
          "answeredQuizId": {
            "type": "string",
            "title": "AnsweredQuizID",
            "example": "fa7f094c-be3a-4c1f-9369-957f7d91d05b"
          },
          "isRetry": {
            "type": "boolean",
            "title": "IsRetry",
            "example": true
          },
          "selectedAnswers": {
            "type": "array",
            "title": "selectedAnswers",
            "example": [
              "are",
              "am"
            ]
          },
          "textAnswer": {
            "type": "string",
            "title": "TextAnswer",
            "example": "jump"
          },
          "isCorrectAnswer": {
            "type": "boolean",
            "title": "IsCorrectAnswer",
            "example": true
          },
          "feedback": {
            "type": "string",
            "title": "Feedback",
            "example": "Great job!"
          },
          "elapsedTime": {
            "type": "number",
            "title": "ElapsedTime",
            "description": "time in seconds",
            "example": 120
          }
        },
        "required": [
          "exerciseId",
          "userId",
          "elapsedTime"
        ]
      },
      "UpdateFeedbackDTO": {
        "type": "object",
        "properties": {
          "feedback": {
            "type": "string",
            "title": "Feedback",
            "example": "Great job!"
          }
        },
        "required": [
          "feedback"
        ]
      },
...
    }
  }
}